cron basic secret

+++ app/api/cron/route.js
@@ -1,14 +1,20 @@
 import connectDB from "@/lib/db/connectDB";
 import payment from "@/lib/models/payment";
 
-export const GET = async ()=> {
+export const GET = async (req) => {
+  const vercelcron = req.headers.get("x-vercel-cron");
+
+  if (!vercelcron) {
+    return new Response("Unauthorized", { status: 401 });
+  }
+
   await connectDB();
 
-  const tenMinutesAgo = new Date(Date.now() - 10 * 60 * 1000)
+  const tenMinutesAgo = new Date(Date.now() - 10 * 60 * 1000);
 
   await payment.deleteMany({
     done: false,
-    createdAt: {$lt: tenMinutesAgo}
-  })
-   return Response.json({ success: true })
-}
+    createdAt: { $lt: tenMinutesAgo },
+  });
+  return Response.json({ success: true });
+};
